create the required nested submodule path and move function and test accordingly
